Carousel set-up for homepage
Some images I've included on the carousel, but please add in any that you both want to as well. We'll need to add in more carousel components per image we use.

Also I followed most of what the wireframe suggested, however I've made some decisions to change the sizing of things. Hope that's okay?